Add net6.0 to the projects
