+++ Directory.Build.targets
@@ -5,14 +5,6 @@
   <Import Project="eng\AndroidX.targets" />
   <Import Project="eng\Microsoft.Extensions.targets" />
 
-  <PropertyGroup>
-    <!--
-      NOTE: Ignore warnings for using Xamarin.iOS in Catalyst
-      error NU1703: 'xunit.core 2.4.1' was resolved as a dependency of 'Microsoft.Maui.TestUtils.DeviceTests', but the dependency is using 'Xamarin.iOS' while 'Microsoft.Maui.TestUtils.DeviceTests' is using 'net6.0-maccatalyst13.5' as its TargetFramework. There might be compatibility issues when MacCatalyst projects depend on Xamarin.iOS projects.
-    -->
-    <NoWarn Condition=" $(TargetFramework.Contains('maccatalyst')) ">$(NoWarn);NU1703</NoWarn>
-  </PropertyGroup>
-
   <!-- Temporarily disable the linker for net6.0-ios and net6.0-maccatalyst until we get machine pools with Big Sur and Xcode 12.5 is stable -->
   <PropertyGroup Condition=" '$(CI)' == 'true' ">
     <PublishTrimmed Condition=" '$(TargetFramework)' == 'net6.0-ios' or '$(TargetFramework)' == 'net6.0-maccatalyst' ">false</PublishTrimmed>
