+++ src/Controls/samples/Controls.Sample/Controls/BordelessEntry/BordelessEntryHandler.cs
@@ -33,7 +33,10 @@ namespace Maui.Controls.Sample.Controls
 #elif WINDOWS
 		public static void MapBorder(BordelessEntryHandler handler, BordelessEntry borderlessEntry)
 		{
-
+		}
+#else
+		public static void MapBorder(BordelessEntryHandler handler, BordelessEntry borderlessEntry)
+		{
 		}
 #endif
 	}
