+++ src/Controls/samples/Controls.Sample/Controls/FocusEffect/FocusRoutingEffect.cs
@@ -1,0 +1,8 @@
+using Microsoft.Maui.Controls;
+
+namespace Maui.Controls.Sample.Controls
+{
+	public class FocusRoutingEffect : RoutingEffect
+	{
+	}
+}
