+++ src/Controls/samples/Controls.Sample/Pages/Core/EffectsPage.xaml
@@ -1,32 +1,32 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<views:BasePage
+﻿<views:BasePage
+
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     x:Class="Maui.Controls.Sample.Pages.EffectsPage"
     xmlns:views="clr-namespace:Maui.Controls.Sample.Pages.Base"
-    xmlns:local="clr-namespace:Maui.Controls.Sample.Pages"
+    xmlns:local="clr-namespace:Maui.Controls.Sample.Controls"
     Title="Alerts">
-    <views:BasePage.Content>
-        <StackLayout
-            Margin="12">
-            <Label
-                Text="Entry With Focus Routing Effect"
-                Style="{StaticResource Headline}"/>
-            <Entry 
-                Text="Alert Simple">
-                <Entry.Effects>
-                    <local:FocusRoutingEffect></local:FocusRoutingEffect>
-                </Entry.Effects>
-            </Entry>
-            <Label
-                Text="Entry With Focus Platform Effect"
-                Style="{StaticResource Headline}"/>
-            <Entry 
-                Text="Alert Simple">
-                <Entry.Effects>
-                    <local:FocusPlatformEffect></local:FocusPlatformEffect>
-                </Entry.Effects>
-            </Entry>
-        </StackLayout>
-    </views:BasePage.Content>
+    <StackLayout
+        Margin="12">
+        <Label
+            Text="Entry With Focus Routing Effect"
+            Style="{StaticResource Headline}"/>
+        <Entry 
+            Text="Alert Simple">
+            <Entry.Effects>
+                <local:FocusRoutingEffect />
+            </Entry.Effects>
+        </Entry>
+        <Label
+            Text="Entry With Focus Platform Effect"
+            Style="{StaticResource Headline}"/>
+        <Entry
+            Text="Alert Simple">
+            <Entry.Effects>
+                <local:FocusPlatformEffect />
+            </Entry.Effects>
+        </Entry>
+    </StackLayout>
+
+
 </views:BasePage>
