+++ src/TestUtils/TestUtils.DeviceTests/src/TestUtils.DeviceTests-net6.csproj
@@ -1,10 +1,9 @@
 ﻿<Project Sdk="MSBuild.Sdk.Extras">
 
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.1;$(MauiPlatforms)</TargetFrameworks>
+    <TargetFrameworks>netstandard2.1;$(MauiPlatformsNoMacCat)</TargetFrameworks>
     <RootNamespace>Microsoft.Maui.TestUtils</RootNamespace>
     <AssemblyName>Microsoft.Maui.TestUtils.DeviceTests</AssemblyName>
-    <NoWarn>$(NoWarn);NU1703</NoWarn>
     <IsPackable>false</IsPackable>
   </PropertyGroup>
